updated typedefs
