+++ packages/legacy/core/App/types/proof-items.ts
@@ -1,0 +1,24 @@
+import { CredentialExchangeRecord } from '@aries-framework/core'
+import { Attribute, Predicate } from '@hyperledger/aries-oca/build/legacy'
+
+export interface ProofCredentialAttributes {
+  altCredentials?: string[]
+  credExchangeRecord?: CredentialExchangeRecord
+  credId: string
+  credDefId?: string
+  schemaId?: string
+  credName: string
+  attributes?: Attribute[]
+}
+
+export interface ProofCredentialPredicates {
+  altCredentials?: string[]
+  credExchangeRecord?: CredentialExchangeRecord
+  credId: string
+  credDefId?: string
+  schemaId?: string
+  credName: string
+  predicates?: Predicate[]
+}
+
+export interface ProofCredentialItems extends ProofCredentialAttributes, ProofCredentialPredicates {}
